refactor: extract zero-run-service, remove identity injection from cli path
- Move startZeroRun() into its own file (zero-run-service.ts)
- Remove buildAgentIdentityPrompt from startRun() (CLI path) — agent
  identity is a platform/integration concern, not CLI
- Default artifactName to "artifact" in startZeroRun() for integration
  callers
- Export resolveByComposeId for use by zero-run-service
