+++ turbo/apps/web/src/lib/run/index.ts
@@ -7,9 +7,9 @@ export {
   validateCheckpoint,
   validateAgentSession,
   startRun,
-  startZeroRun,
   isRunDispatchError,
   type RunDispatchError,
   type StartRunParams,
-  type StartZeroRunParams,
 } from "./run-service";
+
+export { startZeroRun, type StartZeroRunParams } from "./zero-run-service";
