+++ turbo/apps/web/src/lib/run/run-service.ts
@@ -22,7 +22,6 @@ import { orgMetadata } from "../../db/schema/org-metadata";
 import { modelProviders } from "../../db/schema/model-provider";
 import { enqueueRun, drainOrgQueue } from "./run-queue-service";
 import { ORG_SENTINEL_USER_ID } from "../org/org-sentinel";
-import { buildAgentIdentityPrompt } from "../agent-identity";
 import { logger } from "../logger";
 import type { Database } from "../../types/global";
 import type { AgentComposeSnapshot } from "../checkpoint/types";
@@ -488,42 +487,6 @@ export interface StartRunParams {
   checkEnv?: boolean;
 }
 
-/**
- * Params for non-CLI callers (platform UI + integrations).
- *
- * Key differences from StartRunParams:
- * - composeId is always required (all non-CLI callers know the compose)
- * - Only supports composeId + sessionId resolution (no checkpointId, agentComposeVersionId)
- * - No callerOrgId cross-org check (non-CLI callers are already org-scoped)
- */
-export interface StartZeroRunParams {
-  userId: string;
-  prompt: string;
-  composeId: string;
-
-  // Optional — compose resolution
-  sessionId?: string;
-
-  // Optional — forwarded to createRun
-  appendSystemPrompt?: string;
-  disallowedTools?: string[];
-  tools?: string[];
-  settings?: string;
-  conversationId?: string;
-  vars?: Record<string, string>;
-  secrets?: Record<string, string>;
-  artifactName?: string;
-  artifactVersion?: string;
-  memoryName?: string;
-  volumeVersions?: Record<string, string>;
-  scheduleId?: string;
-  callbacks?: Array<{ url: string; secret: string; payload: unknown }>;
-  modelProvider?: string;
-  triggerSource?: TriggerSource;
-  debugNoMockClaude?: boolean;
-  checkEnv?: boolean;
-}
-
 export interface CreateRunResult {
   runId: string;
   status: string;
@@ -938,7 +901,7 @@ async function lookupComposeByVersion(
 /**
  * Resolve compose by composeId → headVersionId.
  */
-async function resolveByComposeId(
+export async function resolveByComposeId(
   composeId: string,
 ): Promise<ResolvedStartRunCompose> {
   const [compose] = await globalThis.services.db
@@ -1032,79 +995,6 @@ async function resolveStartRunCompose(
 }
 
 /**
- * Run entry point for platform UI and integrations (non-CLI callers).
- *
- * Simpler than startRun(): only supports composeId + sessionId resolution,
- * no cross-org check. Resolves compose version + org context internally,
- * injects agent identity, then delegates to createRun().
- *
- * @throws NotFoundError - compose/session not found
- * @throws BadRequestError - compose has no versions
- * @throws ForbiddenError - user cannot access compose
- * @throws Error - dispatch failure
- */
-export async function startZeroRun(
-  params: StartZeroRunParams,
-): Promise<CreateRunResult> {
-  // 1. Resolve compose version: sessionId → resolveBySession, else composeId
-  let resolved: ResolvedStartRunCompose;
-  if (params.sessionId) {
-    const sessionData = await validateAgentSession(
-      params.sessionId,
-      params.userId,
-    );
-    resolved = await resolveByComposeId(sessionData.agentComposeId);
-  } else {
-    resolved = await resolveByComposeId(params.composeId);
-  }
-
-  // 2. Resolve org context
-  const orgData = await getOrgData(resolved.orgId);
-  const orgTier = orgTierSchema.parse(orgData.tier);
-
-  // 3. Inject agent identity into appendSystemPrompt
-  let { appendSystemPrompt } = params;
-  if (resolved.composeId) {
-    const identity = await buildAgentIdentityPrompt(resolved.composeId);
-    if (identity) {
-      appendSystemPrompt = appendSystemPrompt
-        ? `${identity}\n\n${appendSystemPrompt}`
-        : identity;
-    }
-  }
-
-  // 4. Delegate to createRun with fully resolved params
-  return createRun({
-    userId: params.userId,
-    agentComposeVersionId: resolved.agentComposeVersionId,
-    prompt: params.prompt,
-    appendSystemPrompt,
-    disallowedTools: params.disallowedTools,
-    tools: params.tools,
-    settings: params.settings,
-    composeId: resolved.composeId,
-    sessionId: params.sessionId,
-    conversationId: params.conversationId,
-    vars: params.vars,
-    secrets: params.secrets,
-    artifactName: params.artifactName,
-    artifactVersion: params.artifactVersion,
-    memoryName: params.memoryName,
-    volumeVersions: params.volumeVersions,
-    scheduleId: params.scheduleId,
-    callbacks: params.callbacks,
-    agentName: resolved.agentName,
-    modelProvider: params.modelProvider,
-    triggerSource: params.triggerSource,
-    debugNoMockClaude: params.debugNoMockClaude,
-    checkEnv: params.checkEnv,
-    orgSlug: orgData.slug,
-    orgId: resolved.orgId,
-    orgTier,
-  });
-}
-
-/**
  * High-level run entry point for CLI callers.
  *
  * Supports all 4 compose resolution modes (composeId, agentComposeVersionId,
@@ -1137,23 +1027,12 @@ export async function startRun(
   const orgData = await getOrgData(authOrgId);
   const orgTier = orgTierSchema.parse(orgData.tier);
 
-  // 4. Inject agent identity metadata into appendSystemPrompt
-  let { appendSystemPrompt } = params;
-  if (resolved.composeId) {
-    const identity = await buildAgentIdentityPrompt(resolved.composeId);
-    if (identity) {
-      appendSystemPrompt = appendSystemPrompt
-        ? `${identity}\n\n${appendSystemPrompt}`
-        : identity;
-    }
-  }
-
-  // 5. Delegate to createRun with fully resolved params
+  // 4. Delegate to createRun with fully resolved params
   return createRun({
     userId: params.userId,
     agentComposeVersionId: resolved.agentComposeVersionId,
     prompt: params.prompt,
-    appendSystemPrompt,
+    appendSystemPrompt: params.appendSystemPrompt,
     disallowedTools: params.disallowedTools,
     tools: params.tools,
     settings: params.settings,
